+++ shuup/core/order_creator/_source.py
@@ -28,6 +28,7 @@ from shuup.core.models import (
 )
 from shuup.core.pricing import Price, Priceful, TaxfulPrice, TaxlessPrice
 from shuup.core.taxing import should_calculate_taxes_automatically, TaxableItem
+from shuup.core.utils.line_unit_mixin import LineWithUnit
 from shuup.utils.decorators import non_reentrant
 from shuup.utils.i18n import format_money, is_existing_language
 from shuup.utils.money import Money
@@ -554,7 +555,7 @@ class LineSource(Enum):
     DISCOUNT_MODULE = 4
 
 
-class SourceLine(TaxableItem, Priceful):
+class SourceLine(TaxableItem, Priceful, LineWithUnit):
     """
     Line of OrderSource.